[bug] change itemPrice from int to double

+++ models/item.js
@@ -48,7 +48,7 @@ module.exports = (sequelize, DataTypes) => {
         },
       },
       itemPrice: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DOUBLE,
         allowNull: false,
         defaultValue: 0,
         validate: {
@@ -56,6 +56,7 @@ module.exports = (sequelize, DataTypes) => {
             message: 'Item price must be a number',
           },
           isValidPrice(val) {
+            // allows numeric up to two decimals
             const regex = /^[0-9]\d*(((,\d{3}){1})?(\.\d{0,2})?)$/;
             if (!regex.test(val)) {
               throw new Error('Item price is invalid');
